+++ CHANGELOG.md
@@ -34,6 +34,7 @@
 * Add departments to User crate/edit/view pages. Refs UIU-1224.
 * Enable renewal override for Aged to lost items. Refs UIU-1464.
 * Handle Aged to lost items in bulk due date change. Refs UIU-1495.
+* Include Aged to lost in loan details action history. Refs UIU-1803.
 * Add permissions for Departments CRUD. Refs UIU-1778.
 * Add missing permission `departments.collection.get` permission. Fixes UIU-1812.
 * Prevent UI crashing when loading loan with deleted item. Fixes UIU-1819.